Move Editor component to lib
This works towards #12, but does not actually refactor anything - just
keeps the files a bit tidyer.
