+++ frontend/src/lib/editor/Editor.svelte
